+++ ingredient_phrase_tagger/training/translator.py
@@ -1,8 +1,7 @@
 import decimal
 import re
 
-import tokenizer
-import utils
+from ingredient_phrase_tagger.training import utils
 
 
 def translate_row(row):
@@ -20,7 +19,7 @@ def translate_row(row):
     """
     # extract the display name
     display_input = utils.cleanUnicodeFractions(row['input'])
-    tokens = tokenizer.tokenize(display_input)
+    tokens = utils.tokenize(display_input)
 
     labels = _row_to_labels(row)
     label_data = _addPrefixes([(t, _matchUp(t, labels)) for t in tokens])
@@ -97,8 +96,8 @@ def _matchUp(token, labels):
     # Iterate through the labels in descending order of label importance.
     for label_key in ['name', 'unit', 'qty', 'comment', 'range_end']:
         label_value = labels[label_key]
-        if isinstance(label_value, basestring):
-            for n, vt in enumerate(tokenizer.tokenize(label_value)):
+        if isinstance(label_value, str):
+            for n, vt in enumerate(utils.tokenize(label_value)):
                 if utils.normalizeToken(vt) == token:
                     ret.append(label_key.upper())
